Add default LICENSE file
This file should be used as the default LICENSE file for the sources which we provide in
the Big Data Europe project.  Technical partners have been added as the copyright owners
so we won't need to alter the files when contributing to the sources.

The MIT license makes it easy to reuse the project's results as a commercial entity without
adding overly complex constraints.